training/requirements.txt
diffusers[torch]>=0.30.0
peft>=0.12.0
accelerate>=0.33.0
safetensors>=0.4.0
transformers>=4.44.0
Pillow>=10.0.0
torchvision>=0.19.0
